refactor: driver_enum.py /core 아래로 이동
